requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic==2.7.1

# PDF parsing
pymupdf==1.24.3

# Embeddings & Vector Store
openai==1.30.1
faiss-cpu==1.8.0
numpy==1.26.4

# Sparse retrieval
rank-bm25==0.2.2

# HTTP
httpx==0.27.0
aiofiles==23.2.1
python-multipart==0.0.9

# WhatsApp / Twilio
twilio==9.1.0

# Graph
networkx==3.3

# Utilities
tqdm==4.66.4
diskcache==5.6.3
tiktoken==0.7.0
tenacity==8.3.0
